--- app/src/main/java/com/example/beaceful/ui/components/CustomInputField.kt
@@ -1,70 +1,0 @@
-package com.example.beaceful.ui.components
-
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun CustomInputField(
-    @StringRes placeholder: Int,
-    inputText: String,
-    onTextChange: (String) -> Unit,
-    onSent: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        verticalAlignment = Alignment.Top,
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(bottom = 8.dp)
-    ) {
-        OutlinedTextField(
-            value = inputText,
-            onValueChange = {
-                if (it.length <= 120) onTextChange(it)
-            },
-            modifier = Modifier
-                .weight(1f)
-                .fillMaxWidth(),
-            placeholder = { Text(stringResource(placeholder)) },
-            singleLine = false,
-            maxLines = 3,
-            colors = TextFieldDefaults.colors(
-                focusedTextColor = MaterialTheme.colorScheme.primary,
-                unfocusedTextColor = MaterialTheme.colorScheme.primary,
-                focusedPlaceholderColor = MaterialTheme.colorScheme.primary,
-                unfocusedPlaceholderColor = MaterialTheme.colorScheme.primary,
-                focusedContainerColor = MaterialTheme.colorScheme.background,
-                unfocusedContainerColor = MaterialTheme.colorScheme.background,
-                unfocusedIndicatorColor = MaterialTheme.colorScheme.secondary,
-            )
-        )
-
-        Spacer(modifier = Modifier.width(8.dp))
-
-        if (inputText.isNotBlank()) {
-            Button(
-                onClick = onSent,
-                shape = RoundedCornerShape(8.dp),
-                modifier = Modifier.height(56.dp)
-            ) {
-                Text("Đăng")
-            }
-        }
-    }
-}
