+++ app/src/main/java/com/example/beaceful/ui/components/PostDetailsScreen.kt
@@ -14,16 +14,20 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Lock
 import androidx.compose.material.icons.filled.PeopleAlt
 import androidx.compose.material.icons.filled.Public
+import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
@@ -45,9 +49,10 @@ import com.example.beaceful.domain.model.Comment
 import com.example.beaceful.domain.model.DumpDataProvider
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.setValue
-import com.example.beaceful.core.util.formatDateWithHour
+import androidx.compose.ui.res.stringResource
 import com.example.beaceful.domain.model.PostVisibility
 import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
 
 
 @Composable
@@ -65,7 +70,6 @@ fun PostDetailsScreen(
         DumpDataProvider.comments.filter { it.postId == postId }
     }
     var commentText by remember { mutableStateOf("") }
-
     val localComments =
         remember { mutableStateListOf<Comment>().apply { addAll(comments) } }
 
@@ -105,7 +109,7 @@ fun PostDetailsScreen(
                             )
                             Row {
                                 Text(
-                                    formatDateWithHour(post.createdAt),
+                                    post.createdAt.format(DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm")),
                                     color = MaterialTheme.colorScheme.secondary,
                                     style = MaterialTheme.typography.titleSmall
                                 )
@@ -142,20 +146,59 @@ fun PostDetailsScreen(
                 )
             }
             item {
-                CustomInputField (
-                    placeholder = R.string.write_your_comment,
-                    inputText = commentText,
-                    onTextChange = { commentText = it },
-                    onSent = {localComments.add(
-                    Comment(
-                        id = localComments.size + 1,
-                        postId = post.id,
-                        userId = 0,
-                        content = commentText.trim(),
-                        createdAt = LocalDateTime.now()
+                Row(
+                    verticalAlignment = Alignment.Top,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(bottom = 8.dp)
+                ) {
+                    OutlinedTextField(
+                        value = commentText,
+                        onValueChange = {
+                            if (it.length <= 120) commentText = it
+                        },
+                        modifier = Modifier
+                            .weight(1f)
+                            .fillMaxWidth(),
+                        placeholder = { Text(stringResource(R.string.write_your_comment)) },
+                        singleLine = false,
+                        maxLines = 3,
+                        colors = TextFieldDefaults.colors(
+                            focusedTextColor = MaterialTheme.colorScheme.primary,
+                            unfocusedTextColor = MaterialTheme.colorScheme.primary,
+                            focusedPlaceholderColor = MaterialTheme.colorScheme.primary,
+                            unfocusedPlaceholderColor = MaterialTheme.colorScheme.primary,
+                            focusedContainerColor = MaterialTheme.colorScheme.background,
+                            unfocusedContainerColor = MaterialTheme.colorScheme.background,
+                            unfocusedIndicatorColor = MaterialTheme.colorScheme.secondary,
+                        )
+
                     )
-                )
-                    commentText = ""})
+
+                    Spacer(modifier = Modifier.width(8.dp))
+
+                    if (commentText.isNotBlank()) {
+                        Button(
+                            onClick = {
+                                localComments.add(
+                                    Comment(
+                                        id = localComments.size + 1000,  // giả lập ID
+                                        postId = post.id,
+                                        userId = 0,
+                                        content = commentText.trim(),
+                                        createdAt = LocalDateTime.now()
+                                    )
+                                )
+                                commentText = ""
+                            },
+                            enabled = commentText.isNotBlank(),
+                            shape = RoundedCornerShape(8.dp),
+                            modifier = Modifier.height(56.dp)
+                        ) {
+                            Text("Gửi")
+                        }
+                    }
+                }
             }
             if (localComments.size == 0) {
                 item {
@@ -208,7 +251,7 @@ fun CommentCard(comment: Comment) {
                     style = MaterialTheme.typography.titleMedium
                 )
                 Text(
-                    formatDateWithHour(comment.createdAt), color = MaterialTheme.colorScheme.secondary,
+                    comment.createdAt.format(DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm")), color = MaterialTheme.colorScheme.secondary,
                     style = MaterialTheme.typography.titleSmall
                 )
                 Spacer(Modifier.height(6.dp))
