--- app/src/main/java/com/example/beaceful/ui/components/lists/UserList.kt
@@ -1,77 +1,0 @@
-package com.example.beaceful.ui.components.lists
-
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHostController
-import coil3.compose.AsyncImage
-import coil3.request.ImageRequest
-import coil3.request.crossfade
-import com.example.beaceful.R
-import com.example.beaceful.domain.model.User
-import com.example.beaceful.ui.navigation.SingleDoctorProfile
-
-@Composable
-fun UserList(users: List<User>, navController: NavHostController) {
-    LazyRow(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-        contentPadding = PaddingValues(horizontal = 16.dp),
-    ) {
-        items(users) { user ->
-            UserItem(
-                user = user,
-                onClick = { navController.navigate(SingleDoctorProfile.createRoute(user.id)) },
-            )
-        }
-    }
-}
-
-@Composable
-fun UserItem(
-    user: User,
-    onClick: () -> Unit = {},
-) {
-    Column (modifier = Modifier.width(80.dp)) {
-        AsyncImage(
-            model = ImageRequest.Builder(LocalContext.current)
-                .data(user.avatarUrl)
-                .crossfade(true)
-                .build(),
-            placeholder = painterResource(R.drawable.doctor_placeholder_avatar),
-            contentDescription = null,
-            contentScale = ContentScale.Crop,
-            modifier = Modifier
-                .size(80.dp)
-                .clip(CircleShape)
-                .border(
-                    width = 1.dp,
-                    color = MaterialTheme.colorScheme.primary,
-                    shape = CircleShape
-                )
-                .clickable(onClick = onClick)
-        )
-        Spacer(Modifier.height(4.dp))
-        Text(text = user.fullName, textAlign = TextAlign.Center)
-    }
-}
