+++ app/src/main/java/com/example/beaceful/ui/navigation/BeacefulNavHost.kt
@@ -1,24 +1,19 @@
 package com.example.beaceful.ui.navigation
 
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import com.example.beaceful.domain.model.Emotions
 import com.example.beaceful.ui.components.PostDetailsScreen
+import com.example.beaceful.ui.screens.forum.ForumScreen
 import com.example.beaceful.ui.screens.diary.DiaryScreen
-import com.example.beaceful.ui.screens.diary.FullscreenDiaryScreen
-import com.example.beaceful.ui.screens.diary.SelectEmotionScreen
-import com.example.beaceful.ui.screens.diary.WriteDiaryScreen
 import com.example.beaceful.ui.screens.doctor.DoctorScreen
 import com.example.beaceful.ui.screens.doctor.SingleDoctorProfileScreen
-import com.example.beaceful.ui.screens.forum.CommunityScreen
-import com.example.beaceful.ui.screens.forum.ForumScreen
 import com.example.beaceful.ui.screens.home.HomeScreen
 import com.example.beaceful.ui.screens.profile.ProfileScreen
 
@@ -42,16 +37,7 @@ fun BeacefulNavHost(
             DoctorScreen(navController = navController)
         }
         composable(route = Forum.route) {
-            ForumScreen(navController = navController)
-        }
-
-        composable(
-            route = CommunityRoute.route,
-            arguments = listOf(navArgument("communityId") { type = NavType.IntType })
-        ) { backStackEntry ->
-            val communityId = backStackEntry.arguments?.getInt("communityId") ?: return@composable
-            CommunityScreen(navController = navController, communityId = communityId)
-
+            ForumScreen()
         }
         composable(route = Profile.route) {
             ProfileScreen()
@@ -72,32 +58,6 @@ fun BeacefulNavHost(
             PostDetailsScreen(postId = postId)
         }
 
-        composable(
-            route = WriteDiary.route,
-            arguments = listOf(navArgument("emotion") { type = NavType.StringType })
-        ) { backStackEntry ->
-            val emotionArg = backStackEntry.arguments?.getString("emotion")
-            val emotion = runCatching { Emotions.valueOf(emotionArg ?: "") }.getOrNull()
-            if (emotion != null) {
-                WriteDiaryScreen(navController, selectedEmotion = emotion)
-            } else {
-                Text("Emotion không hợp lệ")
-            }
-        }
-        composable(
-            route = WriteDiaryExpand.route
-        ) {
-            FullscreenDiaryScreen(
-                navController = navController
-            )
-        }
-        composable(
-            route = SelectEmotionDiary.route
-        ) {
-            SelectEmotionScreen(
-                navController = navController
-            )
-        }
     }
 }
