+++ app/src/main/java/com/example/beaceful/ui/navigation/BeacefulRoutes.kt
@@ -2,12 +2,15 @@ package com.example.beaceful.ui.navigation
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CollectionsBookmark
+import androidx.compose.material.icons.filled.Create
+import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material.icons.filled.Forum
 import androidx.compose.material.icons.filled.Healing
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.ui.graphics.vector.ImageVector
-import com.example.beaceful.domain.model.Emotions
+import androidx.navigation.NavType
+import androidx.navigation.navArgument
 
 sealed interface BeacefulRoutes {
     val icon: ImageVector
@@ -22,21 +25,6 @@ data object Diary : BeacefulRoutes {
     override val icon = Icons.Filled.CollectionsBookmark
     override val route = "diary"
 }
-data object SelectEmotionDiary : BeacefulRoutes {
-    override val icon = Icons.Filled.CollectionsBookmark
-    override val route = "diary_write"
-}
-data object WriteDiary : BeacefulRoutes {
-    override val icon = Icons.Filled.CollectionsBookmark
-    override val route = "diary_write/{emotion}"
-
-    fun createRoute(emotion: Emotions) : String = "diary_write/${emotion.name}"
-}
-data object WriteDiaryExpand : BeacefulRoutes {
-    override val icon = Icons.Filled.CollectionsBookmark
-    override val route = "diary_write_fullscreen"
-}
-
 data object Doctor : BeacefulRoutes {
     override val icon = Icons.Filled.Healing
     override val route = "doctor"
@@ -44,12 +32,6 @@ data object Doctor : BeacefulRoutes {
 data object Forum : BeacefulRoutes {
     override val icon = Icons.Filled.Forum
     override val route = "forum"
-}
-data object CommunityRoute : BeacefulRoutes {
-    override val icon = Icons.Filled.Forum
-    override val route = "forum_community/{communityId}"
-
-    fun createRoute(communityId: Int) = "forum_community/$communityId"
 }
 data object Profile : BeacefulRoutes {
     override val icon = Icons.Filled.Person
