+++ app/src/main/java/com/example/beaceful/ui/screens/diary/DiaryScreen.kt
@@ -5,22 +5,17 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowForwardIos
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.ArrowBackIosNew
 import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.FloatingActionButtonDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -28,6 +23,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
 import com.example.beaceful.domain.model.Diary
 import com.example.beaceful.domain.model.DumpDataProvider
 import com.example.beaceful.ui.components.cards.DiaryCard
@@ -39,7 +35,7 @@ fun DiaryScreen() {
     val diaries = remember {
         DumpDataProvider.diaries
     }
-    Box {
+    Box() {
         Column {
             Row(
                 verticalAlignment = Alignment.CenterVertically,
@@ -70,16 +66,8 @@ fun DiaryScreen() {
         }
         FloatingActionButton(
             onClick = {},
-            modifier = Modifier
-                .align(Alignment.BottomCenter)
-                .padding(bottom = 18.dp),
-            containerColor = MaterialTheme.colorScheme.primary,
-            shape = RoundedCornerShape(18.dp),
-            contentColor = MaterialTheme.colorScheme.onPrimary,
-            elevation = FloatingActionButtonDefaults.elevation(),
-//            interactionSource = TODO(),
-        )
-        {
+            modifier = Modifier.align(Alignment.BottomCenter).padding(bottom = 24.dp)
+        ) {
             Icon(Icons.Default.Add, contentDescription = null)
         }
     }
@@ -103,9 +91,6 @@ fun DiaryListScreen(
                 onDiaryClick = {
 //                    navController.navigate(SingleDoctorProfile.createRoute(diary.id))
                 })
-        }
-        item {
-            Spacer(Modifier.height(80.dp))
         }
     }
 }
